find pre-existing file and skip download if md5 OK

+++ update-phone.pl
@@ -2,12 +2,12 @@
 # quick script to grab newest video file from phone;
 use strict;
 use warnings;
-use diagnostics;          # useful for debugging;
-use feature 'say';        # beats print;
-use Getopt::Long;         # for parsing command-line options;
-use WWW::Mechanize;       # for reading web pages programmatically;
-use Storable;             # for saving files found by WWW::Mechanize;
-$|++;                     # disable readline buffering for real-time output;
+use diagnostics;       # useful for debugging;
+use feature 'say';     # beats print;
+use Getopt::Long;      # for parsing command-line options;
+use WWW::Mechanize;    # for reading web pages programmatically;
+use Storable;          # for saving files found by WWW::Mechanize;
+$|++;                  # disable readline buffering for real-time output;
 
 my $usage = <<'END';
 update-phone
@@ -26,6 +26,7 @@ Supported options:
 
     -b, --build         # 
     -d, --device         # specify device for cyanogen mod
+    -f, --filename         # specify name for downloaded file (optional)
     -h, --help         # show this usage information
     -v, --verbose        # enable chatty output
 
@@ -33,6 +34,7 @@ END
 
 GetOptions(
     'device|d'       => \my $device,
+    'filename|f'     => \my $filename,
     'build|b'        => \my $build,
     'type|t'         => \my $type,
     'help|h|?|usage' => \my $help,
@@ -48,24 +50,36 @@ $device = 'maguro'  unless $device;
 $type   = 'nightly' unless $type;
 $verbose = 1;      # debugging;
 
-my $start = "http://download.cyanogenmod.com/?device=$device&type=$type";
+download();
 
-my $mech = WWW::Mechanize->new( autocheck => 1 );
-print "Looking up latest CyanogenMod nightly for $device platform... " if $verbose;    # chatty output;
-$mech->get( $start );                                                                  #
-say "done." if $verbose;                                                               # chatty output;
+sub download {     # go grab that file;
 
-$mech->content =~ m/md5sum: (\w{32})/g;                                                # find MD5 for checking;
-my $md5sum = $1;                                                                       # store matched MD5 sum;
+    my $start = "http://download.cyanogenmod.com/?device=$device&type=$type";
 
-my $zip = ( $mech->find_all_links( url_regex => qr/\d{8}\-\w+\-$device\.zip$/ ) )[ 0 ];    # first link on page is most recent;
-my $url = $zip->url_abs;                                                                   # retrieve full URL from object;
+    my $mech = WWW::Mechanize->new( autocheck => 1 );
+    print "Looking up latest CyanogenMod nightly for $device platform... " if $verbose;    # chatty output;
+    $mech->get( $start );                                                                  #
+    say "done." if $verbose;                                                               # chatty output;
 
-say "Fetching image at $url" if $verbose;                                                  # chatty output;
-my $filename = ( split( /\//, $url ) )[ -1 ];                                              # grab name of zipped image file by splitting on URL's /s;
-$mech->get( $url );                                                                        # go grab that url;
-unlink $filename if -e $filename;                                                          # clobber any previous version;
-$mech->save_content( $filename );                                                          # download that file;
+    $mech->content =~ m/md5sum: (\w{32})/g;                                                # find MD5 for checking;
+    my $md5sum = $1;                                                                       # store matched MD5 sum;
 
-say "Verifying file integrity (md5sum: $md5sum)..." if $verbose;                           # chatty output;
-system( 'checkmd5', $md5sum, $filename ) == 0 or die;                                      # make sure the file is what it should be before we flash;
+    my $zip = ( $mech->find_all_links( url_regex => qr/\d{8}\-\w+\-$device\.zip$/ ) )[ 0 ];    # first link on page is most recent;
+    my $url = $zip->url_abs;                                                                   # retrieve full URL from object;
+
+    say "Fetching image at $url" if $verbose;                                                  # chatty output;
+
+    $filename = ( split( /\//, $url ) )[ -1 ]    # grab name of zipped image file by splitting on URL's /s;
+      unless $filename;                          # don't clobber a user-specified filename;
+
+    return 1 if -e $filename and system( 'checkmd5', $md5sum, $filename ) == 0;    # check for pre-existing valid file;
+
+    # continue with download if file isn't already present;
+    unlink $filename if -e $filename;                                              # clobber any previous version;
+
+    $mech->get( $url );                                                            # go grab that url;
+    $mech->save_content( $filename );                                              # download that file;
+
+    say "Verifying file integrity (md5sum: $md5sum)..." if $verbose;               # chatty output;
+    system( 'checkmd5', $md5sum, $filename ) == 0 or die;                          # make sure the file is what it should be before we flash;
+}
